Filtering implemented & attributes supported
Filtering has been added to the suggestions, so it's now possible to load
suggestions from a JSON file or add to JSON directly and filter the results automatically
as the user types.

Attributes are also now supported by use the the "attr" option for each
suggestion in the JSON.

Tests have also been added to test this new functionality.