widthAdjustment option added
A widthAdjustment option has been added so the suggestion-box can be
tweaked.

In addition the showSuggestions() method has been removed as this can be
achieved by using the addSuggestions()/loadSuggestions() and show()
methods, so is essentially duplicating functionailty.

Additional test has been added and test have been changed to deal with the
removal of showSuggestions().