Multiple suggestion boxes supported
Support for multiple suggestion boxes on one page has been added. To make
this work scss/css now uses classes instead of id's and a getId() method
has been added, so the correct suggestion box can be targeted.

In addition the enterKeyAction method has been removed and replaced with
onClick() which allows wasy synchronisation of the click and enter key
actions.

The show method now takes a force paramater, which forces the box to
display, which is useful for debugging and testing.

A reservedKey() function has also been added which allows a convenient
check when implementing custom keyup functionality.