customValues implemented
It's now possible to add custom values to suggestions json that can be
injected in to the suggestion list without being used as part of the filter
pattern.

Fixed an issue that prevents mousover detecting descendents, so not
selecting the suggestion.

Fixed an issue where input was not escaped for filter so was being passed as regex
i.e. /./ would return everything.