Scrollbar support added.
Suggestion box now supports scrollbars and setting of the suggestion box
height.

In addition some tweaking has been done to remove border radius on the
search box when suggestions are shown for a cleaner look.

Suggestion box now remains open when clicked inside the box but not on a
suggestion, this allows scrollbars to be accessed without shutting down
the suggestions, but is actually prefereable behaviour to the previous
closing of the box when non-suggestion were clicked.

Sort, will now sort options when they a loaded if passed. This allows for
more consistent behaviour when overiding default functionality, such as
showing suggestions before user input.